Move ns_string macro to Foundation module
