+++ crates/icrate/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## icrate Unreleased - YYYY-MM-DD
 
+### Changed
+* Moved the `ns_string!` macro to `icrate::Foundation::ns_string`. The old
+  location in the crate root is deprecated.
+
 
 ## icrate 0.0.4 - 2023-07-31
