+++ crates/objc2/CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
-
 ## 0.4.1 - 2023-07-31
 
 ### Added
