+++ crates/test-ui/ui/invalid_ns_string_output.rs
@@ -1,4 +1,4 @@
-use icrate::ns_string;
+use icrate::Foundation::ns_string;
 
 fn main() {
     let _: u8 = ns_string!("abc");
