+++ crates/test-ui/ui/ns_string_not_const.rs
@@ -1,4 +1,4 @@
-use icrate::ns_string;
+use icrate::Foundation::ns_string;
 
 fn main() {
     let s: &str = "abc";
