+++ crates/test-ui/ui/ns_string_output_not_const.rs
@@ -1,5 +1,4 @@
-use icrate::ns_string;
-use icrate::Foundation::NSString;
+use icrate::Foundation::{ns_string, NSString};
 
 fn main() {
     static STRING: &NSString = ns_string!("abc");
