+++ crates/test-ui/ui/ns_string_output_not_const.stderr
@@ -1,4 +1,4 @@
-error[E0015]: cannot call non-const fn `CachedId::<NSString>::get::<[closure@$WORKSPACE/crates/icrate/src/Foundation/macros/ns_string.rs:185:29: 185:31]>` in statics
+error[E0015]: cannot call non-const fn `CachedId::<NSString>::get::<[closure@$WORKSPACE/crates/icrate/src/Foundation/macros/ns_string.rs:189:29: 189:31]>` in statics
  --> ui/ns_string_output_not_const.rs
   |
   |     static STRING: &NSString = ns_string!("abc");
